# prototype/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vesta Report QA API",
    "version": "0.12.0"
  },
  "paths": {
    "/api/v1/config": {
      "get": {
        "summary": "Config",
        "description": "Required scope in api_key mode: reviews:read. Local mode is Vesta-only loopback access.",
        "operationId": "config_api_v1_config_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "QA-Version",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "2026-09-17",
              "title": "Qa-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResource"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "x-required-scope": "reviews:read"
      }
    },
...
    "/api/v1/status": {
      "get": {
        "summary": "Operational Status",
        "description": "Required scope in api_key mode: reviews:read. Local mode is Vesta-only loopback access.",
        "operationId": "operational_status_api_v1_status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "QA-Version",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "2026-09-17",
              "title": "Qa-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "x-required-scope": "reviews:read"
      }
    },
    "/api/v1/diagnostics/openai": {
      "post": {
        "summary": "Openai Connection",
        "description": "Required scope in api_key mode: reviews:read. Local mode is Vesta-only loopback access.",
        "operationId": "openai_connection_api_v1_diagnostics_openai_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "QA-Version",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "2026-09-17",
              "title": "Qa-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "x-required-scope": "reviews:read"
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIErrorDetail": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "param": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Param"
          },
          "field_errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Errors"
          }
        },
        "type": "object",
        "required": [
          "type",
          "code",
          "message",
          "retryable",
          "request_id"
        ],
        "title": "APIErrorDetail"
      },
      "APIErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/APIErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "APIErrorEnvelope"
      },
...
      "ConfigResource": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "demo",
              "openai"
            ],
            "title": "Mode"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version"
          },
          "workflow_version": {
            "type": "string",
            "title": "Workflow Version"
          },
          "policy_status": {
            "type": "string",
            "enum": [
              "supplied_unvalidated",
              "provisional_no_manual"
            ],
            "title": "Policy Status"
          },
          "policy_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Version"
          },
          "ready": {
            "type": "boolean",
            "title": "Ready"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "api_version": {
            "type": "string",
            "title": "Api Version"
          },
          "samples": {
            "items": {
              "$ref": "#/components/schemas/DemoSample"
            },
            "type": "array",
            "title": "Samples"
          }
        },
        "type": "object",
        "required": [
          "mode",
          "model",
          "prompt_version",
          "workflow_version",
          "policy_status",
          "policy_version",
          "ready",
          "tenant_id",
          "api_version",
          "samples"
        ],
        "title": "ConfigResource"
      },
...
      "DemoSample": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "report_text": {
            "type": "string",
            "title": "Report Text"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "report_text"
        ],
        "title": "DemoSample"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}